confirm attendance without pop up to confirm
